[refactor] separate rpgm io
