+++ freesprite/freesprite.vcxproj
@@ -204,6 +204,7 @@
     <ClCompile Include="io_aseprite.cpp" />
     <ClCompile Include="io_gim.cpp" />
     <ClCompile Include="io_piskel.cpp" />
+    <ClCompile Include="io_rpgm.cpp" />
     <ClCompile Include="LayerPalettized.cpp" />
     <ClCompile Include="MainEditorPalettized.cpp" />
     <ClCompile Include="MinecraftBlockPreviewScreen.cpp" />
@@ -363,6 +364,7 @@
     <ClInclude Include="io_base.h" />
     <ClInclude Include="io_gim.h" />
     <ClInclude Include="io_piskel.h" />
+    <ClInclude Include="io_rpgm.h" />
     <ClInclude Include="json\json.hpp" />
     <ClInclude Include="LayerPalettized.h" />
     <ClInclude Include="main.h" />
