+++ freesprite/freesprite.vcxproj.filters
@@ -561,6 +561,9 @@
     <ClCompile Include="CollapsableDraggablePanel.cpp">
       <Filter>Source files\ui_widgets</Filter>
     </ClCompile>
+    <ClCompile Include="io_rpgm.cpp">
+      <Filter>Source files\io</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="globals.h">
@@ -1069,6 +1072,9 @@
     </ClInclude>
     <ClInclude Include="sdl23compat.h">
       <Filter>Header files</Filter>
+    </ClInclude>
+    <ClInclude Include="io_rpgm.h">
+      <Filter>Header files\io</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
